backend/django/drf/terceiro_curso_drf/setup/requirements.txt
asgiref==3.6.0
Django==4.1.6
django-cors-headers==3.13.0
djangorestframework==3.14.0
importlib-metadata==5.2.0
Markdown==3.4.1
pytz==2022.7
sqlparse==0.4.3
zipp==3.11.0
